Finished adding the JPA implementation of the ApplicationDAO
