--- teachus-backend-rdms/src/main/java/dk/teachus/backend/dao/hibernate/applicationContext-hibernate.xml
@@ -1,78 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<beans xmlns="http://www.springframework.org/schema/beans"
-	xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns:jee="http://www.springframework.org/schema/jee"
-	xsi:schemaLocation="
-http://www.springframework.org/schema/beans http://www.springframework.org/schema/beans/spring-beans-3.0.xsd
-http://www.springframework.org/schema/jee http://www.springframework.org/schema/jee/spring-jee-3.0.xsd">
-
-	<!-- Data source -->
-	<jee:jndi-lookup id="dataSource" jndi-name="jdbc/teachus" />
-
-	<!-- LiquiBase Database Schema Evolution Management -->
-	<bean id="liquibase" class="dk.teachus.backend.db.TeachUsSpringLiquibase">
-		<property name="dataSource" ref="dataSource" />
-		<property name="changeLog" value="classpath:db-changelog-master.xml" />
-	</bean>
-
-	<bean id="entityManagerFactory"
-		class="org.springframework.orm.jpa.LocalContainerEntityManagerFactoryBean">
-		<property name="persistenceXmlLocation"
-			value="classpath:/dk/teachus/backend/dao/hibernate/persistence.xml" />
-		<property name="dataSource" ref="dataSource" />
-
-		<property name="jpaVendorAdapter">
-			<bean class="org.springframework.orm.jpa.vendor.HibernateJpaVendorAdapter">
-				<property name="showSql" value="true" />
-				<property name="generateDdl" value="true" />
-				<property name="databasePlatform" value="org.hibernate.dialect.MySQL5Dialect" />
-			</bean>
-		</property>
-	</bean>
-
-<!-- 	<bean id="dynamicDataImport" class="dk.teachus.backend.testdatagenerator.DynamicDataImport"> -->
-<!-- 		<property name="sessionFactory" ref="sessionFactory" /> -->
-<!-- 	</bean> -->
-	<bean id="dynamicDataImport" />
-
-	<!-- +=====+ -->
-	<!-- | DAO | -->
-	<!-- +=====+ -->
-
-	<bean id="personDao" class="dk.teachus.backend.dao.hibernate.HibernatePersonDAO">
-		<property name="sessionFactory">
-			<ref bean="sessionFactory" />
-		</property>
-	</bean>
-
-	<bean id="periodDao" class="dk.teachus.backend.dao.hibernate.HibernatePeriodDAO">
-		<property name="sessionFactory">
-			<ref bean="sessionFactory" />
-		</property>
-	</bean>
-
-	<bean id="bookingDao" class="dk.teachus.backend.dao.hibernate.HibernateBookingDAO">
-		<property name="sessionFactory">
-			<ref bean="sessionFactory" />
-		</property>
-	</bean>
-
-	<bean id="statisticsDao" class="dk.teachus.backend.dao.hibernate.HibernateStatisticsDAO">
-		<property name="sessionFactory">
-			<ref bean="sessionFactory" />
-		</property>
-	</bean>
-
-	<bean id="applicationDao"
-		class="dk.teachus.backend.dao.hibernate.HibernateApplicationDAO">
-		<property name="sessionFactory">
-			<ref bean="sessionFactory" />
-		</property>
-	</bean>
-
-	<bean id="messageDao" class="dk.teachus.backend.dao.hibernate.HibernateMessageDAO">
-		<property name="sessionFactory">
-			<ref bean="sessionFactory" />
-		</property>
-	</bean>
-
-</beans>
