+++ teachus-backend/src/main/java/dk/teachus/backend/applicationContext.xml
@@ -1,11 +1,11 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <beans xmlns="http://www.springframework.org/schema/beans"
-	xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns:jee="http://www.springframework.org/schema/jee"
-	xmlns:tx="http://www.springframework.org/schema/tx"
+	xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns:tx="http://www.springframework.org/schema/tx"
+	xmlns:context="http://www.springframework.org/schema/context"
 	xsi:schemaLocation="
 http://www.springframework.org/schema/beans http://www.springframework.org/schema/beans/spring-beans-3.0.xsd
-http://www.springframework.org/schema/jee http://www.springframework.org/schema/jee/spring-jee-3.0.xsd
-http://www.springframework.org/schema/tx http://www.springframework.org/schema/tx/spring-tx.xsd">
+http://www.springframework.org/schema/tx http://www.springframework.org/schema/tx/spring-tx.xsd
+http://www.springframework.org/schema/context http://www.springframework.org/schema/context/spring-context-3.0.xsd">
 
 	<!-- ===== -->
 	<!-- BEANS -->
@@ -34,35 +34,25 @@ http://www.springframework.org/schema/tx http://www.springframework.org/schema/t
 	</bean>
 
 	<!-- Notification bean -->
-<!-- 	<bean id="notificationBean" class="dk.teachus.backend.bean.impl.VelocityNotificationBean"> -->
-<!-- 		<constructor-arg ref="bookingDao" /> -->
-<!-- 		<constructor-arg ref="personDao" /> -->
-<!-- 		<constructor-arg ref="velocityBean" /> -->
-<!-- 		<constructor-arg ref="messageDao" /> -->
-<!-- 	</bean> -->
+	<!-- <bean id="notificationBean" class="dk.teachus.backend.bean.impl.VelocityNotificationBean"> -->
+	<!-- <constructor-arg ref="bookingDao" /> -->
+	<!-- <constructor-arg ref="personDao" /> -->
+	<!-- <constructor-arg ref="velocityBean" /> -->
+	<!-- <constructor-arg ref="messageDao" /> -->
+	<!-- </bean> -->
 
 	<!-- +====================+ -->
 	<!-- | JPA & Transactions | -->
 	<!-- +====================+ -->
-	
-	<jee:jndi-lookup id="dataSource" jndi-name="jdbc/teachus" />
 
 	<bean id="transactionManager" class="org.springframework.orm.jpa.JpaTransactionManager">
 		<property name="entityManagerFactory" ref="entityManagerFactory" />
 	</bean>
 
-	<bean id="entityManagerFactory" class="org.springframework.orm.jpa.LocalContainerEntityManagerFactoryBean">
-		<property name="dataSource" ref="dataSource" />
-		<property name="persistenceUnitName" value="teachus-default"/>
-		<property name="jpaVendorAdapter">
-			<bean class="org.springframework.orm.jpa.vendor.HibernateJpaVendorAdapter">
-			</bean>
-		</property>
-	</bean>
-	
 	<tx:annotation-driven transaction-manager="transactionManager" />
-	
-	<bean id="persistenceAnnotation" class="org.springframework.orm.jpa.support.PersistenceAnnotationBeanPostProcessor" />
+
+	<bean id="persistenceAnnotation"
+		class="org.springframework.orm.jpa.support.PersistenceAnnotationBeanPostProcessor" />
 
 	<!-- +=====+ -->
 	<!-- | DAO | -->
@@ -70,16 +60,16 @@ http://www.springframework.org/schema/tx http://www.springframework.org/schema/t
 
 	<bean id="applicationDao" class="dk.teachus.backend.dao.jpa.JpaApplicationDAO" />
 
-<!-- 	<bean id="personDao" /> -->
+	<!-- <bean id="personDao" /> -->
 
-<!-- 	<bean id="periodDao" /> -->
+	<!-- <bean id="periodDao" /> -->
 
-<!-- 	<bean id="bookingDao" /> -->
+	<!-- <bean id="bookingDao" /> -->
 
-<!-- 	<bean id="statisticsDao" /> -->
+	<!-- <bean id="statisticsDao" /> -->
 
-<!-- 	<bean id="applicationDao" /> -->
+	<!-- <bean id="applicationDao" /> -->
 
-<!-- 	<bean id="messageDao" /> -->
+	<!-- <bean id="messageDao" /> -->
 
 </beans>
